+++ src/main/java/de/jeff_media/InvUnload/ConfigUpdater.java
@@ -11,7 +11,6 @@ import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-import de.jeff_media.ChestSort.utils.Utils;
 
 public class ConfigUpdater {